requirements.txt
aiogram==3.13.17
python-dotenv